requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
django-restframework==0.0.1
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
